Move ptx_builder/mod.rs into install.rs
